+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - [Tooltips are hidden when clicking on a button][12067].
 - [Fixed bug when clicking header in Table Editor Widget didn't start editing
   it][12064]
+- [When editing cells or header names in Table Editor Widget, `tab` and `enter`
+  keys jumps to next cell/ next row respectively.][12129]
 - [Fixed bugs occurring after renaming project from within graph editor][12106].
 
 [11889]: https://github.com/enso-org/enso/pull/11889
@@ -21,6 +23,7 @@
 [11908]: https://github.com/enso-org/enso/pull/11908
 [12067]: https://github.com/enso-org/enso/pull/12067
 [12064]: https://github.com/enso-org/enso/pull/12064
+[12129]: https://github.com/enso-org/enso/pull/12129
 [12106]: https://github.com/enso-org/enso/pull/12106
 
 #### Enso Standard Library